Fix bigint query

+++ src/util/misc.ts
@@ -17,7 +17,7 @@ export function stringMatches(str: string | number = '', str2: string | number =
 }
 
 export function roboChimpCLRankQuery(userID: bigint) {
-	return `SELECT COUNT(*)
+	return `SELECT COUNT(*)::int
 FROM public.user
 WHERE ((osb_cl_percent + bso_cl_percent) / 2) >= (
 												  SELECT (((COALESCE(osb_cl_percent, 0)) + (COALESCE(bso_cl_percent, 0))) / 2)
